Refining API using generic type restrictions
I realized that by restricting the type of BasicParam<T> to LosslessStringConvertible,
the many subclasses like IntParam would no longer be needed.  Similarly, I could
use a generic Comparable on RangeLimitedParam and get things like String ranges
for "free."  So, this is a big improvement.  Also added more tests, with more to come.