Made the type of argument sequence generic.
The first time I tried to use the argparser, it became obvious I needed to strip off the command name,
and in the process I suddenly wasn't dealing with an array of strings anymore, but an ArraySlice.  That's
one of Swift's quirks... so I made the parseArgs() method generic over any Sequence.  It's more complex
but should be worth it.  Also no code bloat is expected because any given use of the parser will probably
only use one type of sequence against it.

In the process of making this change, I went ahead and made a private struct ArgIterator<T> to track the
verbatim-status, which makes the code look cleaner.